+++ crates/miden-testing/src/executor.rs
@@ -49,7 +49,7 @@ impl<H: SyncHost> CodeExecutor<H> {
     /// To improve the error message quality, convert the returned [`ExecutionError`] into a
     /// [`Report`](miden_objects::assembly::diagnostics::Report).
     pub fn run(self, code: &str) -> Result<Process, ExecutionError> {
-        let assembler = TransactionKernel::testing_assembler().with_debug_mode(true);
+        let assembler = TransactionKernel::with_kernel_library().with_debug_mode(true);
         // TODO: SourceManager.
         let source_manager =
             alloc::sync::Arc::new(miden_objects::assembly::DefaultSourceManager::default())