+++ crates/miden-testing/src/kernel_tests/tx/test_epilogue.rs
@@ -90,10 +90,8 @@ fn test_epilogue() -> anyhow::Result<()> {
         "
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())?;
 
     // The final account is the initial account with the nonce incremented by one.
     let mut final_account = account.clone();
@@ -190,10 +188,8 @@ fn test_compute_output_note_id() -> anyhow::Result<()> {
             "
         );
 
-        let process = &tx_context.execute_code_with_assembler(
-            &code,
-            TransactionKernel::testing_assembler_with_mock_account(),
-        )?;
+        let process = &tx_context
+            .execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())?;
 
         assert_eq!(
             note.assets().commitment(),
@@ -234,10 +230,10 @@ fn test_epilogue_asset_preservation_violation_too_few_input() -> anyhow::Result<
 
     let output_note_1 = NoteBuilder::new(account.id(), rng())
         .add_assets([fungible_asset_1])
-        .build(&TransactionKernel::testing_assembler_with_mock_account())?;
+        .build(&TransactionKernel::with_mock_libraries())?;
     let output_note_2 = NoteBuilder::new(account.id(), rng())
         .add_assets([fungible_asset_2])
-        .build(&TransactionKernel::testing_assembler_with_mock_account())?;
+        .build(&TransactionKernel::with_mock_libraries())?;
 
     let input_note = create_spawn_note(account.id(), vec![&output_note_1, &output_note_2])?;
 
@@ -271,10 +267,8 @@ fn test_epilogue_asset_preservation_violation_too_few_input() -> anyhow::Result<
         "
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
     assert_execution_error!(process, ERR_EPILOGUE_TOTAL_NUMBER_OF_ASSETS_MUST_STAY_THE_SAME);
     Ok(())
 }
@@ -304,13 +298,13 @@ fn test_epilogue_asset_preservation_violation_too_many_fungible_input() -> anyho
 
     let output_note_1 = NoteBuilder::new(account.id(), rng())
         .add_assets([fungible_asset_1])
-        .build(&TransactionKernel::testing_assembler_with_mock_account())?;
+        .build(&TransactionKernel::with_mock_libraries())?;
     let output_note_2 = NoteBuilder::new(account.id(), rng())
         .add_assets([fungible_asset_2])
-        .build(&TransactionKernel::testing_assembler_with_mock_account())?;
+        .build(&TransactionKernel::with_mock_libraries())?;
     let output_note_3 = NoteBuilder::new(account.id(), rng())
         .add_assets([fungible_asset_3])
-        .build(&TransactionKernel::testing_assembler_with_mock_account())?;
+        .build(&TransactionKernel::with_mock_libraries())?;
 
     let input_note = create_spawn_note(
         ACCOUNT_ID_SENDER.try_into()?,
@@ -347,10 +341,8 @@ fn test_epilogue_asset_preservation_violation_too_many_fungible_input() -> anyho
         "
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_EPILOGUE_TOTAL_NUMBER_OF_ASSETS_MUST_STAY_THE_SAME);
     Ok(())
@@ -389,10 +381,8 @@ fn test_block_expiration_height_monotonically_decreases() -> anyhow::Result<()> 
             .replace("{value_2}", &v2.to_string())
             .replace("{min_value}", &v2.min(v1).to_string());
 
-        let process = &tx_context.execute_code_with_assembler(
-            code,
-            TransactionKernel::testing_assembler_with_mock_account(),
-        )?;
+        let process = &tx_context
+            .execute_code_with_assembler(code, TransactionKernel::with_mock_libraries())?;
 
         // Expiry block should be set to transaction's block + the stored expiration delta
         // (which can only decrease, not increase)
@@ -420,10 +410,8 @@ fn test_invalid_expiration_deltas() -> anyhow::Result<()> {
 
     for value in test_values {
         let code = &code_template.replace("{value_1}", &value.to_string());
-        let process = tx_context.execute_code_with_assembler(
-            code,
-            TransactionKernel::testing_assembler_with_mock_account(),
-        );
+        let process =
+            tx_context.execute_code_with_assembler(code, TransactionKernel::with_mock_libraries());
 
         assert_execution_error!(process, ERR_TX_INVALID_EXPIRATION_DELTA);
     }
@@ -453,10 +441,8 @@ fn test_no_expiration_delta_set() -> anyhow::Result<()> {
     end
     ";
 
-    let process = &tx_context.execute_code_with_assembler(
-        code_template,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process = &tx_context
+        .execute_code_with_assembler(code_template, TransactionKernel::with_mock_libraries())?;
 
     // Default value should be equal to u32::max, set in the prologue
     assert_eq!(process.stack.get(EXPIRATION_BLOCK_ELEMENT_IDX).as_int() as u32, u32::MAX);
@@ -496,10 +482,8 @@ fn test_epilogue_increment_nonce_success() -> anyhow::Result<()> {
         "
     );
 
-    tx_context.execute_code_with_assembler(
-        code.as_str(),
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    tx_context
+        .execute_code_with_assembler(code.as_str(), TransactionKernel::with_mock_libraries())?;
     Ok(())
 }
 
@@ -520,7 +504,7 @@ fn epilogue_fails_on_account_state_change_without_nonce_increment() -> anyhow::R
         end
         ";
 
-    let tx_script = ScriptBuilder::with_mock_account_library()?.compile_tx_script(code)?;
+    let tx_script = ScriptBuilder::with_mock_libraries()?.compile_tx_script(code)?;
 
     let result = TransactionContextBuilder::with_noop_auth_account()
         .tx_script(tx_script)