+++ crates/miden-testing/src/kernel_tests/tx/test_faucet.rs
@@ -80,10 +80,7 @@ fn test_mint_fungible_asset_succeeds() -> anyhow::Result<()> {
     );
 
     let process = &tx_context
-        .execute_code_with_assembler(
-            &code,
-            TransactionKernel::testing_assembler_with_mock_account(),
-        )
+        .execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())
         .unwrap();
 
     let expected_final_storage_amount = FUNGIBLE_FAUCET_INITIAL_BALANCE + FUNGIBLE_ASSET_AMOUNT;
@@ -118,7 +115,7 @@ fn mint_fungible_asset_fails_on_non_faucet_account() -> anyhow::Result<()> {
       ",
         asset = Word::from(FungibleAsset::mock(50))
     );
-    let tx_script = ScriptBuilder::with_mock_account_library()?.compile_tx_script(code)?;
+    let tx_script = ScriptBuilder::with_mock_libraries()?.compile_tx_script(code)?;
 
     let result = TransactionContextBuilder::new(account)
         .tx_script(tx_script)
@@ -151,10 +148,8 @@ fn test_mint_fungible_asset_inconsistent_faucet_id() -> anyhow::Result<()> {
         asset = Word::from(FungibleAsset::mock(5))
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_FUNGIBLE_ASSET_FAUCET_IS_NOT_ORIGIN);
     Ok(())
@@ -173,7 +168,7 @@ fn test_mint_fungible_asset_fails_saturate_max_amount() -> anyhow::Result<()> {
     ",
         asset = Word::from(FungibleAsset::mock(FungibleAsset::MAX_AMOUNT))
     );
-    let tx_script = ScriptBuilder::with_mock_account_library()?.compile_tx_script(code)?;
+    let tx_script = ScriptBuilder::with_mock_libraries()?.compile_tx_script(code)?;
 
     let result = TransactionContextBuilder::with_fungible_faucet(
         FungibleAsset::mock_issuer().into(),
@@ -243,10 +238,7 @@ fn test_mint_non_fungible_asset_succeeds() -> anyhow::Result<()> {
     );
 
     tx_context
-        .execute_code_with_assembler(
-            &code,
-            TransactionKernel::testing_assembler_with_mock_account(),
-        )
+        .execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())
         .unwrap();
     Ok(())
 }
@@ -273,10 +265,8 @@ fn test_mint_non_fungible_asset_fails_inconsistent_faucet_id() -> anyhow::Result
         non_fungible_asset = word_to_masm_push_string(&non_fungible_asset.into())
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_NON_FUNGIBLE_ASSET_FAUCET_IS_NOT_ORIGIN);
     Ok(())
@@ -298,7 +288,7 @@ fn mint_non_fungible_asset_fails_on_non_faucet_account() -> anyhow::Result<()> {
       ",
         asset = Word::from(FungibleAsset::mock(50))
     );
-    let tx_script = ScriptBuilder::with_mock_account_library()?.compile_tx_script(code)?;
+    let tx_script = ScriptBuilder::with_mock_libraries()?.compile_tx_script(code)?;
 
     let result = TransactionContextBuilder::new(account)
         .tx_script(tx_script)
@@ -331,10 +321,8 @@ fn test_mint_non_fungible_asset_fails_asset_already_exists() -> anyhow::Result<(
         non_fungible_asset = word_to_masm_push_string(&non_fungible_asset.into())
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_FAUCET_NON_FUNGIBLE_ASSET_ALREADY_ISSUED);
     Ok(())
@@ -391,10 +379,7 @@ fn test_burn_fungible_asset_succeeds() -> anyhow::Result<()> {
     );
 
     let process = &tx_context
-        .execute_code_with_assembler(
-            &code,
-            TransactionKernel::testing_assembler_with_mock_account(),
-        )
+        .execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())
         .unwrap();
 
     let expected_final_storage_amount = FUNGIBLE_FAUCET_INITIAL_BALANCE - FUNGIBLE_ASSET_AMOUNT;
@@ -429,7 +414,7 @@ fn burn_fungible_asset_fails_on_non_faucet_account() -> anyhow::Result<()> {
       ",
         asset = Word::from(FungibleAsset::mock(50))
     );
-    let tx_script = ScriptBuilder::with_mock_account_library()?.compile_tx_script(code)?;
+    let tx_script = ScriptBuilder::with_mock_libraries()?.compile_tx_script(code)?;
 
     let result = TransactionContextBuilder::new(account)
         .tx_script(tx_script)
@@ -465,10 +450,8 @@ fn test_burn_fungible_asset_inconsistent_faucet_id() -> anyhow::Result<()> {
         suffix = faucet_id.suffix(),
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_FUNGIBLE_ASSET_FAUCET_IS_NOT_ORIGIN);
     Ok(())
@@ -500,10 +483,8 @@ fn test_burn_fungible_asset_insufficient_input_amount() -> anyhow::Result<()> {
         saturating_amount = CONSUMED_ASSET_1_AMOUNT + 1
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_VAULT_FUNGIBLE_ASSET_AMOUNT_LESS_THAN_AMOUNT_TO_WITHDRAW);
     Ok(())
@@ -578,10 +559,7 @@ fn test_burn_non_fungible_asset_succeeds() -> anyhow::Result<()> {
     );
 
     tx_context
-        .execute_code_with_assembler(
-            &code,
-            TransactionKernel::testing_assembler_with_mock_account(),
-        )
+        .execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())
         .unwrap();
     Ok(())
 }
@@ -609,10 +587,8 @@ fn test_burn_non_fungible_asset_fails_does_not_exist() -> anyhow::Result<()> {
         non_fungible_asset = word_to_masm_push_string(&non_fungible_asset_burnt.into())
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_FAUCET_NON_FUNGIBLE_ASSET_TO_BURN_NOT_FOUND);
     Ok(())
@@ -634,7 +610,7 @@ fn burn_non_fungible_asset_fails_on_non_faucet_account() -> anyhow::Result<()> {
       ",
         asset = Word::from(FungibleAsset::mock(50))
     );
-    let tx_script = ScriptBuilder::with_mock_account_library()?.compile_tx_script(code)?;
+    let tx_script = ScriptBuilder::with_mock_libraries()?.compile_tx_script(code)?;
 
     let result = TransactionContextBuilder::new(account)
         .tx_script(tx_script)
@@ -670,10 +646,8 @@ fn test_burn_non_fungible_asset_fails_inconsistent_faucet_id() -> anyhow::Result
         non_fungible_asset = word_to_masm_push_string(&non_fungible_asset_burnt.into())
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_FAUCET_NON_FUNGIBLE_ASSET_TO_BURN_NOT_FOUND);
     Ok(())
@@ -721,10 +695,7 @@ fn test_is_non_fungible_asset_issued_succeeds() -> anyhow::Result<()> {
     );
 
     tx_context
-        .execute_code_with_assembler(
-            &code,
-            TransactionKernel::testing_assembler_with_mock_account(),
-        )
+        .execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())
         .unwrap();
     Ok(())
 }
@@ -760,10 +731,7 @@ fn test_get_total_issuance_succeeds() -> anyhow::Result<()> {
     );
 
     tx_context
-        .execute_code_with_assembler(
-            &code,
-            TransactionKernel::testing_assembler_with_mock_account(),
-        )
+        .execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())
         .unwrap();
     Ok(())
 }
@@ -779,7 +747,7 @@ fn setup_non_faucet_account() -> anyhow::Result<Account> {
     let faucet_component = AccountComponent::compile(
         "export.::miden::faucet::mint
          export.::miden::faucet::burn",
-        TransactionKernel::testing_assembler_with_mock_account(),
+        TransactionKernel::with_mock_libraries(),
         vec![],
     )?
     .with_supported_type(AccountType::RegularAccountUpdatableCode);