+++ crates/miden-testing/src/kernel_tests/tx/test_fpi.rs
@@ -76,7 +76,7 @@ fn test_fpi_memory() -> anyhow::Result<()> {
 
     let foreign_account_component = AccountComponent::compile(
         foreign_account_code_source,
-        TransactionKernel::testing_assembler(),
+        TransactionKernel::with_kernel_library(),
         storage_slots.clone(),
     )?
     .with_supports_all_types();
@@ -322,14 +322,14 @@ fn test_fpi_memory_two_accounts() -> anyhow::Result<()> {
 
     let foreign_account_component_1 = AccountComponent::compile(
         foreign_account_code_source_1,
-        TransactionKernel::testing_assembler(),
+        TransactionKernel::with_kernel_library(),
         storage_slots_1.clone(),
     )?
     .with_supports_all_types();
 
     let foreign_account_component_2 = AccountComponent::compile(
         foreign_account_code_source_2,
-        TransactionKernel::testing_assembler(),
+        TransactionKernel::with_kernel_library(),
         storage_slots_2.clone(),
     )?
     .with_supports_all_types();
@@ -529,7 +529,7 @@ fn test_fpi_execute_foreign_procedure() -> anyhow::Result<()> {
 
     let foreign_account_component = AccountComponent::compile(
         foreign_account_code_source,
-        TransactionKernel::testing_assembler(),
+        TransactionKernel::with_kernel_library(),
         storage_slots,
     )?
     .with_supports_all_types();
@@ -689,7 +689,7 @@ fn test_nested_fpi_cyclic_invocation() -> anyhow::Result<()> {
 
     let second_foreign_account_component = AccountComponent::compile(
         second_foreign_account_code_source,
-        TransactionKernel::testing_assembler(),
+        TransactionKernel::with_kernel_library(),
         storage_slots,
     )?
     .with_supports_all_types();
@@ -747,7 +747,7 @@ fn test_nested_fpi_cyclic_invocation() -> anyhow::Result<()> {
 
     let first_foreign_account_component = AccountComponent::compile(
         first_foreign_account_code_source,
-        TransactionKernel::testing_assembler(),
+        TransactionKernel::with_kernel_library(),
         storage_slots,
     )?
     .with_supports_all_types();
@@ -886,7 +886,7 @@ fn test_nested_fpi_stack_overflow() {
             let storage_slots = vec![AccountStorage::mock_item_0().slot];
             let last_foreign_account_component = AccountComponent::compile(
                 last_foreign_account_code_source,
-                TransactionKernel::testing_assembler(),
+                TransactionKernel::with_kernel_library(),
                 storage_slots,
             )
             .unwrap()
@@ -933,7 +933,7 @@ fn test_nested_fpi_stack_overflow() {
 
                 let foreign_account_component = AccountComponent::compile(
                     foreign_account_code_source,
-                    TransactionKernel::testing_assembler(),
+                    TransactionKernel::with_kernel_library(),
                     vec![],
                 )
                 .unwrap()
@@ -1050,7 +1050,7 @@ fn test_nested_fpi_native_account_invocation() -> anyhow::Result<()> {
 
     let foreign_account_component = AccountComponent::compile(
         foreign_account_code_source,
-        TransactionKernel::testing_assembler(),
+        TransactionKernel::with_kernel_library(),
         vec![],
     )?
     .with_supports_all_types();
@@ -1145,7 +1145,7 @@ fn test_fpi_stale_account() -> anyhow::Result<()> {
 
     let foreign_account_component = AccountComponent::compile(
         foreign_account_code_source,
-        TransactionKernel::testing_assembler(),
+        TransactionKernel::with_kernel_library(),
         vec![AccountStorage::mock_item_0().slot],
     )?
     .with_supports_all_types();