+++ crates/miden-testing/src/kernel_tests/tx/test_output_note.rs
@@ -156,7 +156,7 @@ fn test_get_asset_info() -> anyhow::Result<()> {
         assets_number_1 = output_note_1.assets().num_assets(),
     );
 
-    let tx_script = ScriptBuilder::with_kernel_library()?.compile_tx_script(tx_script_src)?;
+    let tx_script = ScriptBuilder::default().compile_tx_script(tx_script_src)?;
 
     let tx_context = mock_chain
         .build_tx_context(account.id(), &[], &[])?
@@ -229,7 +229,7 @@ fn test_get_recipient_and_metadata() -> anyhow::Result<()> {
         METADATA = word_to_masm_push_string(&output_note.metadata().into()),
     );
 
-    let tx_script = ScriptBuilder::with_kernel_library()?.compile_tx_script(tx_script_src)?;
+    let tx_script = ScriptBuilder::default().compile_tx_script(tx_script_src)?;
 
     let tx_context = mock_chain
         .build_tx_context(account.id(), &[], &[])?
@@ -332,7 +332,7 @@ fn test_get_assets() -> anyhow::Result<()> {
         check_note_2 = check_assets_code(2, 8, &p2id_note_2_assets),
     );
 
-    let tx_script = ScriptBuilder::with_kernel_library()?.compile_tx_script(tx_script_src)?;
+    let tx_script = ScriptBuilder::default().compile_tx_script(tx_script_src)?;
 
     let tx_context = mock_chain
         .build_tx_context(account.id(), &[], &[])?