+++ crates/miden-testing/src/kernel_tests/tx/test_tx.rs
@@ -224,10 +224,8 @@ fn test_create_note() -> anyhow::Result<()> {
         tag = tag,
     );
 
-    let process = &tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process =
+        &tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())?;
 
     assert_eq!(
         process.get_kernel_mem_word(NUM_OUTPUT_NOTES_PTR),
@@ -276,7 +274,7 @@ fn test_create_note_with_invalid_tag() -> anyhow::Result<()> {
         tx_context
             .execute_code_with_assembler(
                 &note_creation_script(invalid_tag),
-                TransactionKernel::testing_assembler()
+                TransactionKernel::with_kernel_library()
             )
             .is_err()
     );
@@ -285,7 +283,7 @@ fn test_create_note_with_invalid_tag() -> anyhow::Result<()> {
         tx_context
             .execute_code_with_assembler(
                 &note_creation_script(valid_tag),
-                TransactionKernel::testing_assembler()
+                TransactionKernel::with_kernel_library()
             )
             .is_ok()
     );
@@ -352,10 +350,8 @@ fn test_create_note_too_many_notes() -> anyhow::Result<()> {
         aux = Felt::ZERO,
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_TX_NUMBER_OF_OUTPUT_NOTES_EXCEEDS_LIMIT);
     Ok(())
@@ -513,10 +509,8 @@ fn test_get_output_notes_commitment() -> anyhow::Result<()> {
         )),
     );
 
-    let process = &tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process =
+        &tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())?;
 
     assert_eq!(
         process.get_kernel_mem_word(NUM_OUTPUT_NOTES_PTR),
@@ -591,10 +585,8 @@ fn test_create_note_and_add_asset() -> anyhow::Result<()> {
         asset = word_to_masm_push_string(&asset),
     );
 
-    let process = &tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process =
+        &tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())?;
 
     assert_eq!(
         process.get_kernel_mem_word(OUTPUT_NOTE_SECTION_OFFSET + OUTPUT_NOTE_ASSETS_OFFSET),
@@ -676,10 +668,8 @@ fn test_create_note_and_add_multiple_assets() -> anyhow::Result<()> {
         nft = word_to_masm_push_string(&non_fungible_asset_encoded),
     );
 
-    let process = &tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process =
+        &tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())?;
 
     assert_eq!(
         process.get_kernel_mem_word(OUTPUT_NOTE_SECTION_OFFSET + OUTPUT_NOTE_ASSETS_OFFSET),
@@ -756,10 +746,8 @@ fn test_create_note_and_add_same_nft_twice() -> anyhow::Result<()> {
         nft = word_to_masm_push_string(&encoded),
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_NON_FUNGIBLE_ASSET_ALREADY_EXISTS);
     Ok(())
@@ -857,10 +845,8 @@ fn test_build_recipient_hash() -> anyhow::Result<()> {
         aux = aux,
     );
 
-    let process = &tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process =
+        &tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())?;
 
     assert_eq!(
         process.get_kernel_mem_word(NUM_OUTPUT_NOTES_PTR),
@@ -903,10 +889,8 @@ fn test_block_procedures() -> anyhow::Result<()> {
         end
         ";
 
-    let process = &tx_context.execute_code_with_assembler(
-        code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process =
+        &tx_context.execute_code_with_assembler(code, TransactionKernel::with_mock_libraries())?;
 
     assert_eq!(
         process.stack.get_word(0),