+++ crates/miden-testing/src/tx_context/context.rs
@@ -131,11 +131,12 @@ impl TransactionContext {
         .execute_program(program)
     }
 
-    /// Executes arbitrary code with a testing assembler ([TransactionKernel::testing_assembler()]).
+    /// Executes arbitrary code with a kernel library assembler
+    /// ([`TransactionKernel::with_kernel_library`]).
     ///
     /// For more information, see the docs for [TransactionContext::execute_code_with_assembler()].
     pub fn execute_code(&self, code: &str) -> Result<Process, ExecutionError> {
-        let assembler = TransactionKernel::testing_assembler();
+        let assembler = TransactionKernel::with_kernel_library();
         self.execute_code_with_assembler(code, assembler)
     }