+++ crates/miden-testing/tests/scripts/send_note.rs
@@ -44,10 +44,7 @@ fn test_send_note_script_basic_wallet() -> anyhow::Result<()> {
         Default::default(),
     )?;
     let assets = NoteAssets::new(vec![sent_asset]).unwrap();
-    let note_script = ScriptBuilder::with_kernel_library()
-        .unwrap()
-        .compile_note_script("begin nop end")
-        .unwrap();
+    let note_script = ScriptBuilder::default().compile_note_script("begin nop end").unwrap();
     let serial_num = RpoRandomCoin::new(Word::from([1, 2, 3, 4u32])).draw_word();
     let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
 
@@ -110,10 +107,7 @@ fn test_send_note_script_basic_fungible_faucet() -> anyhow::Result<()> {
     let assets = NoteAssets::new(vec![Asset::Fungible(
         FungibleAsset::new(sender_basic_fungible_faucet_account.id(), 10).unwrap(),
     )])?;
-    let note_script = ScriptBuilder::with_kernel_library()
-        .unwrap()
-        .compile_note_script("begin nop end")
-        .unwrap();
+    let note_script = ScriptBuilder::default().compile_note_script("begin nop end").unwrap();
     let serial_num = RpoRandomCoin::new(Word::from([1, 2, 3, 4u32])).draw_word();
     let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());